start moving email functions to user-emails file
