+++ api/src/application/user/user-service.js
@@ -221,28 +221,6 @@ module.exports = (bookshelf, emailService, clientService, renderTemplate, RedisA
         html: template,
       });
     },
-
-    async sendVerificationEmail(email, query, client) {
-      const base = config('/baseUrl');
-      const viewContext = {
-        url: `${base}/user/email-verify`,
-        appName: client.get('client_name'),
-      };
-
-      let template = await themeService.renderThemedTemplate(query.client_id, 'forgot-password-email', viewContext);
-
-      if (!template) {
-        template = await renderTemplate('email-verify-email', viewContext, {
-          layout: 'email',
-        });
-      }
-
-      emailService.send({
-        to: email,
-        subject: 'Verify your email address',
-        html: template,
-      });
-    }
   };
 
   return self;
